// connections/connections/src/app/store/action-type/group-message-action.types.ts
export enum GroupMessageActionTypes {
  GET_GROUP_MESSAGE = '[GROUP_PAGE] GET GROUP_MESSAGE',
  GET_GROUP__MESSAGE_SUCCESSFUL = '[GROUP_PAGE] GET GROUP_MESSAGE SUCCESSFUL',
  GET_GROUP__MESSAGE_FAILED = '[GROUP_PAGE] GET GROUP_MESSAGE FAILED',

  SEND_GROUP_MESSAGE = '[GROUP_PAGE] SEND GROUP_MESSAGE',
  SEND_GROUP__MESSAGE_SUCCESSFUL = '[GROUP_PAGE] SEND GROUP_MESSAGE SUCCESSFUL',
  SEND_GROUP__MESSAGE_FAILED = '[GROUP_PAGE] SEND GROUP_MESSAGE FAILED',

  UPDATE_GROUP_MESSAGE = '[GROUP_PAGE] UPDATE GROUP MESSAGE',
  UPDATE_GROUP_MESSAGE_SUCCESSFUL = '[GROUP_PAGE] UPDATE GROUP MESSAGE SUCCESSFUL',
  UPDATE_GROUP_MESSAGE_FAILED = '[GROUP_PAGE] UPDATE GROUP MESSAGE FAILED'
}